debbug problem with use history
